clap turns 3

+++ {{project-name}}/src/lib.rs
@@ -1,4 +1,4 @@
-pub fn double(x: i32) -> i32 {
+pub fn double(x: i64) -> i64 {
     x * 2
 }
